Bug fix + graphical polish
Fix a bug kindly spotted by Don @CarlosMejZ: if a new simulation is
requested with too many cereals, and another simulation was already
running, after printing the error message the GUI returns to run the
previous one. Apparently the while(True) was not killed by the button.
Fixed by defining a proper logical field in the class and switching it
on and off when appropriate.

The graphical polish, suggested by @SamueleGiuli, consists in defining
tighter xlim and ylim, as to avoid having the PBCs looking weird (now
they just more or less simulate the sticking to the edge of the bowl).

Co-authored-by: SamueleGiuli <samuele.giuli@gmail.com>